% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureDat.R
\docType{data}
\name{featureDat}
\alias{featureDat}
\title{Protein-intrinsic feature data in MAPD}
\format{
A data frame, including all human proteins (row) and their features.
}
\usage{
data(featureDat)
}
\description{
The collection and processing of the data are fully described in our manuscript.
}
\examples{
data(featureDat)
head(featureDat)
}
\keyword{datasets}
