% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPD.CV.R
\name{MAPD.CV}
\alias{MAPD.CV}
\title{Cross-validation.}
\usage{
MAPD.CV(model, fold = 20, metric = "ROC", out = "curve")
}
\arguments{
\item{model}{A caret model returned by the train function.}

\item{fold}{An integer, specifying the fold for cross validation.}

\item{metric}{A character, One of ROC and PRC.}

\item{out}{A character, specifying the type of output.
Should be one of curve, AUC, F1, and Pred.}
}
\value{
The output depends the `out` parameter. Could be a ROC/PRC object
or AUC value or prediction scores collected from the cross-validation.
}
\description{
Cross-validation.
}
\note{
This function merges predicted probability from each recursion (each-fold of samples)
and evaluate the performance based on prediction scores for all samples.
}
