% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPD.R
\docType{data}
\name{MAPD}
\alias{MAPD}
\title{A pre-trained random forest model using caret}
\format{
A caret::train object.
}
\usage{
data(MAPD)
}
\description{
caret::train object
}
\examples{
data(MAPD)
}
\keyword{datasets}
