% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRC.plot.R
\name{PRC.plot}
\alias{PRC.plot}
\title{Draw Precision-Recall Curve using ggplot2.}
\usage{
PRC.plot(prcList)
}
\arguments{
\item{prcList}{One or a list of PRROC::pr.curve object.}
}
\value{
A ggplot instance.
}
\description{
Draw Precision-Recall Curve using ggplot2.
}
