% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinsForTrain.R
\docType{data}
\name{ProteinsForTrain}
\alias{ProteinsForTrain}
\title{Highly-degradable and lowly-degradable kinases for training the model}
\format{
A factor including kinases (names of the factor) and their label (high or low).
}
\usage{
data(ProteinsForTrain)
}
\description{
Highly-degradable and lowly-degradable kinases for training the model
}
\examples{
data(ProteinsForTrain)
head(ProteinsForTrain)
}
\keyword{datasets}
