% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.plot.R
\name{ROC.plot}
\alias{ROC.plot}
\title{Draw ROC (Receiver operating characteristic) Curve using ggplot2.}
\usage{
ROC.plot(rocList)
}
\arguments{
\item{rocList}{One or a list of pROC::roc object.}
}
\value{
A ggplot instance.
}
\description{
Draw ROC (Receiver operating characteristic) Curve using ggplot2.
}
